# perf: improve performance for rules