@gajus 2022-03-26
:tada: This PR is included in version 38.0.7 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/eslint-plugin-jsdoc/v/38.0.7)
- [GitHub release](https://github.com/gajus/eslint-plugin-jsdoc/releases/tag/v38.0.7)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: